This commit was manufactured by cvs2svn to create tag 'SQUID_2_4_STABLE6'.

+++ src/repl/heap/store_heap_replacement.c
@@ -43,7 +43,6 @@
 
 #include "squid.h"
 #include "heap.h"
-#include "store_heap_replacement.h"
 
 /*
  * Key generation function to implement the LFU-DA policy (Least
@@ -65,7 +64,7 @@
  * for this to become a problem. (estimation is 10^8 cache
  * turnarounds)
  */
-heap_key
+heap_key 
 HeapKeyGen_StoreEntry_LFUDA(void *entry, double age)
 {
     StoreEntry *e = entry;
@@ -78,8 +77,8 @@ HeapKeyGen_StoreEntry_LFUDA(void *entry, double age)
     else
 	tie = 1.0 - exp((double) (e->lastref - squid_curtime) / 86400.0);
     key = age + (double) e->refcount - tie;
-    debug(81, 3) ("HeapKeyGen_StoreEntry_LFUDA: %s refcnt=%d lastref=%ld age=%f tie=%f -> %f\n",
-	storeKeyText(e->hash.key), (int)e->refcount, e->lastref, age, tie, key);
+    debug(81, 3) ("HeapKeyGen_StoreEntry_LFUDA: %s refcnt=%ld lastref=%ld age=%f tie=%f -> %f\n",
+	storeKeyText(e->hash.key), e->refcount, e->lastref, age, tie, key);
     if (e->mem_obj && e->mem_obj->url)
 	debug(81, 3) ("HeapKeyGen_StoreEntry_LFUDA: url=%s\n",
 	    e->mem_obj->url);
@@ -106,7 +105,7 @@ HeapKeyGen_StoreEntry_LFUDA(void *entry, double age)
  * for this to become a problem. (estimation is 10^8 cache
  * turnarounds)
  */
-heap_key
+heap_key 
 HeapKeyGen_StoreEntry_GDSF(void *entry, double age)
 {
     StoreEntry *e = entry;
@@ -114,8 +113,8 @@ HeapKeyGen_StoreEntry_GDSF(void *entry, double age)
     double size = e->swap_file_sz ? (double) e->swap_file_sz : 1.0;
     double tie = (e->lastref > 1) ? (1.0 / e->lastref) : 1.0;
     key = age + ((double) e->refcount / size) - tie;
-    debug(81, 3) ("HeapKeyGen_StoreEntry_GDSF: %s size=%f refcnt=%d lastref=%ld age=%f tie=%f -> %f\n",
-	storeKeyText(e->hash.key), size, (int)e->refcount, e->lastref, age, tie, key);
+    debug(81, 3) ("HeapKeyGen_StoreEntry_GDSF: %s size=%f refcnt=%ld lastref=%ld age=%f tie=%f -> %f\n",
+	storeKeyText(e->hash.key), size, e->refcount, e->lastref, age, tie, key);
     if (e->mem_obj && e->mem_obj->url)
 	debug(81, 3) ("HeapKeyGen_StoreEntry_GDSF: url=%s\n",
 	    e->mem_obj->url);
@@ -129,7 +128,7 @@ HeapKeyGen_StoreEntry_GDSF(void *entry, double age)
  * Don't use it unless you are trying to compare performance among
  * heap-based replacement policies...
  */
-heap_key
+heap_key 
 HeapKeyGen_StoreEntry_LRU(void *entry, double age)
 {
     StoreEntry *e = entry;
